Remove unused scope variable. Fixes bug #589.
When we created a directive for the merge button in pr #583 we set the state parameters to the scope of the controller in order to pass the variables to the template. Setting the user variable did overwrite the user variable of the controller. Later on we changed the structure of the directive so it is not necessary to set the varibles to the scope anymore. This commite removes setting those variables to the scope of the controller. This fixes #589.